Refactor: refactor test
