+++ packages/binary/valuetransfer/transfer/output_test.go
@@ -1,4 +1,4 @@
-package transferoutput
+package transfer
 
 import (
 	"testing"
