+++ packages/binary/valuetransfer/transfer/outputmetadata.go
@@ -1,4 +1,4 @@
-package transferoutput
+package transfer
 
 import (
 	"sync"
